Feature: NRF consumer support oauth2
