+++ go.mod
@@ -9,7 +9,7 @@ require (
 	github.com/free5gc/aper v1.0.4
 	github.com/free5gc/nas v1.1.0
 	github.com/free5gc/ngap v1.0.6
-	github.com/free5gc/openapi v1.0.7-0.20230802173229-2b3ded4db293
+	github.com/free5gc/openapi v1.0.7-0.20231112094355-a96c3450377e
 	github.com/free5gc/pfcp v1.0.6
 	github.com/free5gc/util v1.0.5-0.20231001095115-433858e5be94
 	github.com/gin-gonic/gin v1.9.1
